+++ nodeconductor/structure/models.py
@@ -494,13 +494,14 @@ class Project(core_models.DescribableMixin,
         return self.name
 
     def get_users(self, role=None):
-        users = get_user_model().objects.filter(
+        query = Q(
             projectpermission__project=self,
             projectpermission__is_active=True,
         )
         if role:
-            users = users.filter(projectpermission__role=role)
-        return users
+            query = query & Q(projectpermission__role=role)
+
+        return get_user_model().objects.filter(query)
 
     def __str__(self):
         return '%(name)s | %(customer)s' % {